add release mechanism to the packet
